+++ src/app/BasePage.js
@@ -14,7 +14,7 @@ import { ListeTachesEtudiant } from './pages/ListeTachesEtudiant';
 import { MyPage } from './pages/MyPage';
 import { TacheDetails } from './pages/TacheDetails';
 import { Tuteurs } from './pages/Tuteurs';
-import { getAllGroups } from './services/Api';
+import { getUsersListByRole } from './services/Users';
 
 
 export default function BasePage() {
@@ -22,14 +22,14 @@ export default function BasePage() {
 
   useEffect(() => {
     location.pathname.includes('/groupe/') && localStorage.setItem('groupeNb', location.pathname.substr(location.pathname.length - 4));
-    getAllGroups()
-          .then( response => {
-            console.log(response.data);
-           
-          })
+    
+    getUsersListByRole('002')
+          .then( response => console.log(response))
           .catch(() => {
-           
+
           });
+;
+
   }, []) // [] - is required if you need only one call
   // https://reactjs.org/docs/hooks-reference.html#useeffect
